Add levenshtein_myers_8x16 implementation

+++ inc/levenshtein_myers.hpp
@@ -8,7 +8,7 @@ struct Myers8x16Input {
   const char *q_wrd;
   int q_wrd_len;
   const char *d_wrds[16];
-  uint32_t d_wrd_lens[16];
+  uint8_t d_wrd_lens[16];
 };
 
 struct Myers16x8Input {
@@ -32,7 +32,7 @@ struct Myers64x2Input {
   uint64_t d_wrd_lens[2];
 };
 
-std::array<uint32_t, 16> levenshtein_myers_8x16(const Myers8x16Input &input);
+std::array<uint8_t, 16> levenshtein_myers_8x16(const Myers8x16Input &input);
 std::array<uint16_t, 8> levenshtein_myers_16x8(const Myers16x8Input &input);
 std::array<uint32_t, 4> levenshtein_myers_32x4(const Myers32x4Input &input);
 std::array<uint64_t, 2> levenshtein_myers_64x2(const Myers64x2Input &input);
